# Cache `eslint` in pre-commit script (makes `git commit` 5x faster)